# Conflicts:
#	IGraphics/Controls/Test/TestMultiPathControl.h

+++ IGraphics/IGraphicsPathBase.h
@@ -38,7 +38,7 @@ public:
     float height = bitmap.H() / bitmap.GetDrawScale();
     
     PathTransformSave();
-    PathTransformTranslate((float) destCtrX, (float) destCtrY);
+    PathTransformTranslate(destCtrX, destCtrY);
     PathTransformRotate((float) angle);
     DrawBitmap(bitmap, IRECT(-width * 0.5f, - height * 0.5f, width * 0.5f, height * 0.5f), 0, 0, pBlend);
     PathTransformRestore();
